deprecate old test_utils/synthetic/make_processing_configs, replace with new version
