+++ tests/synthetic/test_synthetic_driver.py
@@ -3,8 +3,7 @@ from aurora.pipelines.process_mth5 import process_mth5_run
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test1_h5
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test2_h5
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test12rr_h5
-from aurora.test_utils.synthetic.make_processing_configs_new import \
-    create_test_run_config
+from aurora.test_utils.synthetic.make_processing_configs import create_test_run_config
 from aurora.test_utils.synthetic.processing_helpers import process_sythetic_data
 from aurora.tf_kernel.dataset import DatasetDefinition
 from aurora.tf_kernel.helpers import extract_run_summaries_from_mth5s
